backend/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication and Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Configuration and Environment
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP Clients
httpx>=0.25.0
requests>=2.31.0

# Data Validation
email-validator>=2.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Redis for caching/sessions
redis>=5.0.0

# Optional: PostgreSQL driver
psycopg2-binary>=2.9.0

# AI Integration
anthropic>=0.7.0
openai>=1.3.0